Update README.md
Clarifying country support, as there is no worldwide unified framework for this. The owner/developer is based in the US and currently only speaks English. The developer is open to other submitting PR requests. There are many integrations that are specific to specific countries as such, it is normal for integrations to be region specific.